Fix logging crashing with segfault on arm32
Likely caused by architecture dependent difference in types of the
values passed to printf. Now explicitly cast to long to match the format
specifier.

+++ src/rogue/Logging.cpp
@@ -106,7 +106,7 @@ void rogue::Logging::intLog(uint32_t level, const char* fmt, va_list args) {
     char buffer[1000];
     vsnprintf(buffer, sizeof(buffer), fmt, args);
     gettimeofday(&tme, NULL);
-    printf("%l" PRIi32 ".%06l" PRIi32 ":%s: %s\n", tme.tv_sec, tme.tv_usec, name_.c_str(), buffer);
+    printf("%l" PRIi32 ".%06l" PRIi32 ":%s: %s\n", (long)tme.tv_sec, (long)tme.tv_usec, name_.c_str(), buffer);
 }
 
 void rogue::Logging::log(uint32_t level, const char* fmt, ...) {
